requirements.txt
streamlit==1.51.0
pillow==11.3.0
borax==4.1.3
google-genai==1.56.0

